+++ Changelog.md
@@ -31,6 +31,7 @@
 - Added the `Stockable` interface to the `Product` and `MasterProductVariant` models
 - Added the `backorder` field to products and product variants 
 - BC: Added the `?CheckoutSubject` return type to the `getCart()` method of the `Checkout` interface
+- BC: The unused `$config` parameter has been removed from the `RequestStore` checkout driver constructor
 - Fixed possible null return type on Billpayer::getName() when is_organization is true but the company name is null
 
 ## 3.x Series